requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.3.4-cp37-cp37m-win32.whl
allure-pytest